Refactor : single party dto 패키지 내 response, request 패키지 만듦
party - dto - single 패키지 내에 response, request 패키지를 만들었습니다.
