+++ src/main/java/nbbang/com/nbbang/domain/party/controller/PartyController.java
@@ -12,8 +12,12 @@ import nbbang.com.nbbang.domain.member.domain.Member;
 import nbbang.com.nbbang.domain.member.service.MemberService;
 import nbbang.com.nbbang.domain.party.domain.Party;
 import nbbang.com.nbbang.domain.party.dto.*;
-import nbbang.com.nbbang.domain.party.dto.many.PartyFindRequestDto;
 import nbbang.com.nbbang.domain.party.dto.single.*;
+import nbbang.com.nbbang.domain.party.dto.single.request.PartyChangeGoalNumberRequestDto;
+import nbbang.com.nbbang.domain.party.dto.single.request.PartyRequestDto;
+import nbbang.com.nbbang.domain.party.dto.single.request.PartyStatusChangeRequestDto;
+import nbbang.com.nbbang.domain.party.dto.single.response.PartyIdResponseDto;
+import nbbang.com.nbbang.domain.party.dto.single.response.PartyReadResponseDto;
 import nbbang.com.nbbang.domain.party.service.PartyService;
 import nbbang.com.nbbang.global.error.exception.CustomIllegalArgumentException;
 import nbbang.com.nbbang.global.response.DefaultResponse;
@@ -24,7 +28,6 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
